% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{dat}
\alias{dat}
\title{simulated data for demonstrating the features of interep}
\format{Each data consists of six components: e, z, x, y, coef and index; index shows the location of the true coefficients used to generate y.}
\usage{
data("dat")
}
\description{
Simulated data for demonstrating the features of interep.
}
\examples{
data("dat")
}
\keyword{datasets}
