% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty.R
\name{penalty}
\alias{penalty}
\title{This function gives the penalty functions}
\usage{
penalty(x, n, p, q, beta, lam1, pmethod, p1, lam2)
}
\arguments{
\item{x}{matrix of covariates.}

\item{n}{the sample size.}

\item{p}{the number of predictors.}

\item{q}{the number of environment factors.}

\item{beta}{the coefficient vector.}

\item{lam1}{the tuning parameter lambda1 for individual penalty.}

\item{pmethod}{the penalization method. "mixed" refers to MCP penalty to individual main effects and group MCP penalty to interactions; "individual" means MCP penalty to all effects.}

\item{p1}{the number of gene factors.}

\item{lam2}{the tuning parameter lambda2 for group penalty.}
}
\value{
\item{E}{the penalty function.}
}
\description{
This function gives the penalty functions
}
