% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat}
\alias{reformat}
\title{This function reformats the longituinal data (wide to long format)}
\usage{
reformat(k, y, x)
}
\arguments{
\item{k}{the number of repeated measurement.}

\item{y}{the longitudinal response.}

\item{x}{a matrix of predictors, consisting of omics and environment factors, as well as their interactions. In the case study,
the omics measurements are lipidomics data.}
}
\value{
the long format of the longitudinal data.
}
\description{
This function reformats the longituinal data (wide to long format)
}
