% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreU.R
\name{ScoreU}
\alias{ScoreU}
\title{This function obtains the score function and its derivative}
\usage{
ScoreU(n, k, y, x, p, response = "continuous", beta, Rhat)
}
\arguments{
\item{n}{the sample size.}

\item{k}{the number of repeated measurement.}

\item{y}{the longitudinal response.}

\item{x}{a matrix of predictors, consisting of omics and environment factors, as well as their interactions. In the case study,
the omics measurements are lipidomics data.}

\item{p}{number of predictors.}

\item{response}{type of the longitudinal response, the default is continuous.}

\item{beta}{a coefficient vector.}

\item{Rhat}{the correlation matrix.}
}
\value{
the score function and its derivative.
}
\description{
This function obtains the score function and its derivative
}
