% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reformat}
\alias{reformat}
\title{This function changes the format of the longitudinal data from wide format to long format}
\usage{
reformat(y, x)
}
\arguments{
\item{y}{the longitudinal response.}

\item{x}{a matrix of predictors, consisting of omics and environment factors, as well as their interactions. In
the case study, the omics measurements are lipidomics data.}
}
\description{
This function changes the format of the longitudinal data from wide format to long format
}
\examples{
data("dat")
y=dat$y
x=dat$x
reformat(y,x)
}
